# backend/test_api.py
#!/usr/bin/env python3
"""
Test script for the FastAPI RAG Chatbot API
"""

import pytest
from fastapi.testclient import TestClient
import sys
import os
sys.path.append(os.path.dirname(os.path.abspath(__file__)))
from api import app

# Create test client
client = TestClient(app)

def test_health_endpoint():
    """Test the health endpoint"""
    response = client.get("/health")
    assert response.status_code == 200
    data = response.json()
    assert "status" in data
    assert data["status"] in ["healthy", "unhealthy"]


def test_root_endpoint():
    """Test the root endpoint"""
    response = client.get("/")
    assert response.status_code == 200
    data = response.json()
    assert data["status"] == "ok"
    assert "message" in data


def test_system_info_endpoint():
    """Test the system info endpoint"""
    response = client.get("/system-info")
    assert response.status_code == 200
    data = response.json()
    assert "collection_name" in data
    assert "model_name" in data
    assert "use_gemini" in data


def test_validate_query_endpoint():
    """Test the validate query endpoint"""
    # Test a valid query
    response = client.get("/validate-query", params={"question": "What is ROS 2?"})
    assert response.status_code == 200
    data = response.json()
    assert data["is_valid"] is True
    assert data["question"] == "What is ROS 2?"

    # Test an invalid query (too short) - this will trigger FastAPI validation
    # which returns 422, so we test with a query that's too short but passes FastAPI validation
    response = client.get("/validate-query", params={"question": "A"})  # Too short for business validation
    # Note: FastAPI's Query validation requires min_length=3, so this will return 422
    # This is expected behavior - the API validates at the parameter level
    assert response.status_code == 422  # FastAPI validation error for too short query

    # Test with a query that passes FastAPI validation but fails business validation
    # Actually, the business validation in agent.py also checks for min length of 3,
    # so we need to test with a different type of invalid query
    response = client.get("/validate-query", params={"question": "ab"})  # Still too short
    assert response.status_code == 422


def test_query_endpoint():
    """Test the main chat endpoint with a simple query"""
    # Test a basic query
    response = client.post("/chat", json={
        "question": "What is the purpose of this system?",
        "top_k": 3,
        "min_score": 0.3
    })
    assert response.status_code == 200
    data = response.json()
    assert "content" in data
    assert "sources" in data
    assert "grounded" in data
    assert isinstance(data["sources"], list)
    assert isinstance(data["grounded"], bool)


def test_query_endpoint_with_selected_text():
    """Test the chat endpoint with selected text scope"""
    # Test with selected text
    response = client.post("/chat", json={
        "question": "What does this text explain?",
        "selected_text": "ROS 2 is a set of software libraries and tools that help you build robot applications.",
        "top_k": 3,
        "min_score": 0.3
    })
    assert response.status_code == 200
    data = response.json()
    assert "content" in data
    assert "sources" in data


def main():
    """Run all tests"""
    print("Running API tests...")

    test_health_endpoint()
    print("[OK] Health endpoint test passed")

    test_root_endpoint()
    print("[OK] Root endpoint test passed")

    test_system_info_endpoint()
    print("[OK] System info endpoint test passed")

    test_validate_query_endpoint()
    print("[OK] Validate query endpoint test passed")

    test_query_endpoint()
    print("[OK] Query endpoint test passed")

    test_query_endpoint_with_selected_text()
    print("[OK] Query endpoint with selected text test passed")

    print("\nAll tests passed! [OK]")


if __name__ == "__main__":
    main()